+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.16] - 2022-01-26
+### Changed
+- UncookedConnector - Encoding defaults to None.
+
+### Added
+- UncookedConnector - supports file_mode and encoding.
+
 ## [0.0.15] - 2021-12-02
 ### Fixed
 - SqlAlchemyDatabaseConnector needs to call it's own callable optional args so it can supply the declarive base argument. i.e. non-simple callable. Created a mechanism for Connectors to reserve this right so Connect() will leave these callables.